+++ tests/integrations/opentelemetry/test_span_processor.py
@@ -3,19 +3,13 @@ from datetime import datetime, timezone
 from unittest import mock
 from unittest.mock import MagicMock
 
-import pytest
 from opentelemetry.trace import SpanKind, SpanContext, Status, StatusCode
 
 from sentry_sdk.integrations.opentelemetry.span_processor import (
     SentrySpanProcessor,
     link_trace_context_to_error_event,
 )
-from sentry_sdk.integrations.opentelemetry.utils import (
-    extract_span_data,
-    is_sentry_span,
-    span_data_for_db_query,
-    span_data_for_http_method,
-)
+from sentry_sdk.integrations.opentelemetry.utils import is_sentry_span
 from sentry_sdk.scope import Scope
 from sentry_sdk.tracing import Span, Transaction
 from sentry_sdk.tracing_utils import extract_sentrytrace_data
@@ -215,28 +209,6 @@ def test_update_span_with_otel_data_http_method():
     assert sentry_span._data["http.user_agent"] == "curl/7.64.1"
     assert sentry_span._data["net.peer.name"] == "example.com"
     assert sentry_span._data["http.target"] == "/"
-
-
-@pytest.mark.parametrize(
-    "otel_status, expected_status",
-    [
-        pytest.param(Status(StatusCode.UNSET), None, id="unset"),
-        pytest.param(Status(StatusCode.OK), "ok", id="ok"),
-        pytest.param(Status(StatusCode.ERROR), "internal_error", id="error"),
-    ],
-)
-def test_update_span_with_otel_status(otel_status, expected_status):
-    sentry_span = Span()
-
-    otel_span = MagicMock()
-    otel_span.name = "Test OTel Span"
-    otel_span.kind = SpanKind.INTERNAL
-    otel_span.status = otel_status
-
-    span_processor = SentrySpanProcessor()
-    span_processor._update_span_with_otel_status(sentry_span, otel_span)
-
-    assert sentry_span.get_trace_context().get("status") == expected_status
 
 
 def test_update_span_with_otel_data_http_method2():
@@ -611,357 +583,3 @@ def test_pruning_old_spans_on_end():
     span_processor.on_end(otel_span)
     assert sorted(list(span_processor.otel_span_map.keys())) == ["111111111abcdef"]
     assert sorted(list(span_processor.open_spans.values())) == [{"111111111abcdef"}]
-
-
-def test_span_data_for_db_query():
-    otel_span = MagicMock()
-    otel_span.name = "OTel Span"
-    otel_span.attributes = {}
-
-    op, description, status, http_status = span_data_for_db_query(otel_span)
-    assert op == "db"
-    assert description == "OTel Span"
-    assert status is None
-    assert http_status is None
-
-    otel_span.attributes = {"db.statement": "SELECT * FROM table;"}
-
-    op, description, status, http_status = span_data_for_db_query(otel_span)
-    assert op == "db"
-    assert description == "SELECT * FROM table;"
-    assert status is None
-    assert http_status is None
-
-
-@pytest.mark.parametrize(
-    "kind, status, attributes, expected",
-    [
-        (
-            SpanKind.CLIENT,
-            None,  # None means unknown error
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-            },
-            {
-                "op": "http.client",
-                "description": "POST /some/route",
-                "status": "unknown_error",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            Status(StatusCode.UNSET),  # Unset defaults to OK
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            None,
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-                "http.status_code": 502,  # Take this status in case of None status
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "internal_error",
-                "http_status_code": 502,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            Status(StatusCode.UNSET),
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-                "http.status_code": 502,  # Take this status in case of UNSET status
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "internal_error",
-                "http_status_code": 502,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            None,
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-                "http.status_code": 502,
-                "http.response.status_code": 503,  # this takes precedence over deprecated http.status_code
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "unavailable",
-                "http_status_code": 503,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            Status(StatusCode.UNSET),
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-                "http.status_code": 502,
-                "http.response.status_code": 503,  # this takes precedence over deprecated http.status_code
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "unavailable",
-                "http_status_code": 503,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            Status(StatusCode.OK),  # OK status is taken right away
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            Status(StatusCode.OK),  # OK status is taken right away
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-                "http.response.status_code": 200,
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "ok",
-                "http_status_code": 200,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            Status(
-                StatusCode.ERROR
-            ),  # Error status without description gets the http status from attributes
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-                "http.response.status_code": 401,
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "unauthenticated",
-                "http_status_code": 401,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            Status(
-                StatusCode.ERROR, "I'm a teapot"
-            ),  # Error status with unknown description is an unknown error
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-                "http.response.status_code": 418,
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "unknown_error",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.SERVER,
-            Status(
-                StatusCode.ERROR, "unimplemented"
-            ),  # Error status with known description is taken (grpc errors)
-            {
-                "http.method": "POST",
-                "http.route": "/some/route",
-            },
-            {
-                "op": "http.server",
-                "description": "POST /some/route",
-                "status": "unimplemented",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.CLIENT,
-            Status(StatusCode.OK),
-            {
-                "http.method": "GET",
-                "http.target": None,  # no location for description
-                "net.peer.name": None,
-                "http.url": None,
-            },
-            {
-                "op": "http.client",
-                "description": "GET",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.CLIENT,
-            Status(StatusCode.OK),
-            {
-                "http.method": "GET",
-                "http.target": "/target",  # this can be the location in the description
-            },
-            {
-                "op": "http.client",
-                "description": "GET /target",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.CLIENT,
-            Status(StatusCode.OK),
-            {
-                "http.method": "GET",
-                "net.peer.name": "example.com",  # this can be the location in the description
-            },
-            {
-                "op": "http.client",
-                "description": "GET example.com",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.CLIENT,
-            Status(StatusCode.OK),
-            {
-                "http.method": "GET",
-                "http.target": "/target",  # target takes precedence over net.peer.name
-                "net.peer.name": "example.com",
-            },
-            {
-                "op": "http.client",
-                "description": "GET /target",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            SpanKind.CLIENT,
-            Status(StatusCode.OK),
-            {
-                "http.method": "GET",
-                "http.url": "https://username:secretpwd@example.com/bla/?secret=123&anothersecret=456",  # sensitive data is stripped
-            },
-            {
-                "op": "http.client",
-                "description": "GET https://example.com/bla/",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-    ],
-)
-def test_span_data_for_http_method(kind, status, attributes, expected):
-    otel_span = MagicMock()
-    otel_span.kind = kind
-    otel_span.status = status
-    otel_span.attributes = attributes
-
-    op, description, status, http_status_code = span_data_for_http_method(otel_span)
-    result = {
-        "op": op,
-        "description": description,
-        "status": status,
-        "http_status_code": http_status_code,
-    }
-    assert result == expected
-
-
-@pytest.mark.parametrize(
-    "name, status, attributes, expected",
-    [
-        (
-            "OTel Span Blank",
-            Status(StatusCode.UNSET),  # Unset defaults to OK
-            {},
-            {
-                "op": "OTel Span Blank",
-                "description": "OTel Span Blank",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            "OTel Span RPC",
-            Status(StatusCode.UNSET),  # Unset defaults to OK
-            {
-                "rpc.service": "myservice.EchoService",
-            },
-            {
-                "op": "rpc",
-                "description": "OTel Span RPC",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            "OTel Span Messaging",
-            Status(StatusCode.UNSET),  # Unset defaults to OK
-            {
-                "messaging.system": "rabbitmq",
-            },
-            {
-                "op": "message",
-                "description": "OTel Span Messaging",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-        (
-            "OTel Span FaaS",
-            Status(StatusCode.UNSET),  # Unset defaults to OK
-            {
-                "faas.trigger": "pubsub",
-            },
-            {
-                "op": "pubsub",
-                "description": "OTel Span FaaS",
-                "status": "ok",
-                "http_status_code": None,
-            },
-        ),
-    ],
-)
-def test_extract_span_data(name, status, attributes, expected):
-    otel_span = MagicMock()
-    otel_span.name = name
-    otel_span.status = Status(StatusCode.UNSET)
-    otel_span.attributes = attributes
-
-    op, description, status, http_status_code = extract_span_data(otel_span)
-    result = {
-        "op": op,
-        "description": description,
-        "status": status,
-        "http_status_code": http_status_code,
-    }
-    assert result == expected
